sntj_api/pyproject.toml
[project]
name = "sntj-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "backoff>=2.2.1",
    "fastapi[standard]>=0.115.0",
    "httpx>=0.27.0",
    "loguru>=0.7.3",
    "pyarrow>=16.0.0",
    "pydantic-settings>=2.8.1",
    "pydantic>=2.4.0",
    "pytest>=8.1.0",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.2",
    "sqlalchemy>=2.0.0",
    "uvicorn>=0.23.0",
    "azure-identity>=1.20.0",
    "pyodbc>=5.2.0",
]
